pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cotpressure"
version = "0.1.0"
description = "CoT Pressure: Preference Learning with Text-Based Lie Detectors"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "CoT Pressure Team"}
]
dependencies = [
    # Core ML
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "datasets>=2.18.0",
    "accelerate>=0.27.0",
    "peft>=0.10.0",
    "bitsandbytes>=0.43.0",

    # Training frameworks
    "trl>=0.8.0",
    "unsloth @ git+https://github.com/unslothai/unsloth.git",

    # Inference
    "vllm>=0.4.0",

    # Detector training
    "scikit-learn>=1.4.0",

    # Utilities
    "pyyaml>=6.0",
    "wandb>=0.16.0",
    "tqdm>=4.66.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[project.scripts]
cotpressure = "cotpressure.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]
